Fold main.cpp into AutoPilotBot.cpp and move drawing code to StrategyManager
