+++ src/starterbot/AutoPilotBot.h
@@ -10,21 +10,26 @@ class AutoPilotBot : public EventReceiver {
 private:
 	StrategyManager m_strategyManager;
 
+	int m_gameCount;
+
+public:
+	// Creates an instance of AutoPilotBot, connects to the BWAPI client, and plays games.
+	static void runBot();
+
 protected:
 	virtual void notifyMembers(const bw::Event& event) override;
 
 	virtual void onStart() override;
-	virtual void onDraw() override;
 	virtual void onEnd(bool isWinner) override;
 
 private:
-	// Draws bounding boxes around each unit plus a line indicating what the target of the
-	// bot's current command is.
-	void drawUnitBoxes();
-	void drawCommands();
+	// We don't want people to construct AutoPilotBot except by calling runBot().
+	AutoPilotBot() = default;
 
-	// Draws bars above each unit representing the percentage of health, shields, or
-	// resources that the unit has.
-	void drawHealthBars();
-	void drawHealthBar(bw::Unit unit, double ratio, bw::Color color, int yOffset);
+	// Tries to connect to the BWAPI client repeatedly, waiting if the connection failed.
+	// Upon connecting, this calls playGame() repeatedly until the client disconnects.
+	void initLoop();
+	// Waits for a game to start, and then takes every event from BWAPI as they come and
+	// dispatches them to notifyReceiver() until the game stops or the client disconnects.
+	void playGame();
 };
