+++ src/starterbot/StrategyManager.cpp
@@ -24,6 +24,13 @@ void StrategyManager::onFrame() {
     }
 }
 
+void StrategyManager::onDraw() {
+    // Draw some useful information about each unit, namely commands and health.
+    drawUnitBoxes();
+    drawCommands();
+    drawHealthBars();
+}
+
 void StrategyManager::onProtossFrame() {
     // First, we make sure none of our buildings are idle by training new units.
     m_productionManager.idleTrainRequests(bw::UnitTypes::Protoss_Probe, false);
@@ -120,5 +127,87 @@ void StrategyManager::onZergFrame() {
     // Like the other races, send a scout once we have a decent number of workers.
     if (workerCount >= 9 && m_scoutManager.countScouts(bw::UnitTypes::Zerg_Drone) == 0) {
         m_scoutManager.addScout(bw::UnitTypes::Zerg_Drone);
+    }
+}
+
+void StrategyManager::drawUnitBoxes() {
+    for (bw::Unit unit : g_game->getAllUnits()) {
+        bw::Position topLeft(unit->getLeft(), unit->getTop());
+        bw::Position bottomRight(unit->getRight(), unit->getBottom());
+
+        g_game->drawBoxMap(topLeft, bottomRight, bw::Colors::White);
+    }
+}
+
+void StrategyManager::drawCommands() {
+    for (bw::Unit unit : g_self->getUnits()) {
+        const bw::UnitCommand& command = unit->getLastCommand();
+
+        // If the previous command had a ground position target, draw it in green.
+        if (command.getTargetPosition() != bw::Positions::None) {
+            g_game->drawLineMap(unit->getPosition(),
+                command.getTargetPosition(), bw::Colors::Green);
+        }
+
+        // If the previous command had a unit target, draw it in white.
+        if (command.getTarget() != nullptr) {
+            g_game->drawLineMap(unit->getPosition(),
+                command.getTarget()->getPosition(), bw::Colors::White);
+        }
+    }
+}
+
+void StrategyManager::drawHealthBars() {
+    for (bw::Unit unit : g_game->getAllUnits()) {
+        // If the unit is a resource, draw the remaining resources in cyan.
+        if (unit->getType().isResourceContainer() && unit->getInitialResources() > 0) {
+            double mineralRatio = (double)unit->getResources() / (double)unit->getInitialResources();
+            drawHealthBar(unit, mineralRatio, bw::Colors::Cyan, 0);
+        }
+
+        // If the unit has health, then draw the health in green, orange, or red,
+        // corresponding to how damaged the unit is.
+        if (unit->getType().maxHitPoints() > 0) {
+            double hpRatio = (double)unit->getHitPoints() / (double)unit->getType().maxHitPoints();
+
+            bw::Color hpColor;
+            if (hpRatio < 0.33) {
+                hpColor = bw::Colors::Red;
+            } else if (hpRatio < 0.66) {
+                hpColor = bw::Colors::Orange;
+            } else {
+                hpColor = bw::Colors::Green;
+            }
+
+            drawHealthBar(unit, hpRatio, hpColor, 0);
+        }
+
+        // If the unit has shields, draw those as well in blue.
+        if (unit->getType().maxShields() > 0) {
+            double shieldRatio = (double)unit->getShields() / (double)unit->getType().maxShields();
+            drawHealthBar(unit, shieldRatio, bw::Colors::Blue, -3);
+        }
+    }
+}
+
+void StrategyManager::drawHealthBar(bw::Unit unit, double ratio, bw::Color color, int yOffset) {
+    bw::Position pos = unit->getPosition();
+    int unitTop = pos.y - unit->getType().dimensionUp();
+
+    // Calculate the dimensions for the bar background and length.
+    int left = pos.x - unit->getType().dimensionLeft();
+    int right = pos.x + unit->getType().dimensionRight();
+    int top = unitTop + yOffset - 10;
+    int bottom = unitTop + yOffset - 6;
+
+    int bar = (int)((right - left) * ratio + left);
+
+    // Draw the background of the bar, the bar itself, and the tick marks inside the bar.
+    g_game->drawBoxMap(bw::Position(left, top), bw::Position(right, bottom), bw::Colors::Grey, true);
+    g_game->drawBoxMap(bw::Position(left, top), bw::Position(bar, bottom), color, true);
+    g_game->drawBoxMap(bw::Position(left, top), bw::Position(right, bottom), bw::Colors::Black, false);
+
+    for (int i = left; i < right - 1; i += 3) {
+        g_game->drawLineMap(bw::Position(i, top), bw::Position(i, bottom), bw::Colors::Black);
     }
 }
