+++ src/starterbot/StrategyManager.h
@@ -24,6 +24,7 @@ protected:
     virtual void notifyMembers(const bw::Event& event) override;
 
     virtual void onFrame() override;
+    virtual void onDraw() override;
 
 private:
     // Since the strategy for each race is substantially different, a different strategy
@@ -31,4 +32,14 @@ private:
     void onProtossFrame();
     void onTerranFrame();
     void onZergFrame();
+
+    // Draws bounding boxes around each unit plus a line indicating what the target of the
+    // bot's current command is.
+    void drawUnitBoxes();
+    void drawCommands();
+
+    // Draws bars above each unit representing the percentage of health, shields, or
+    // resources that the unit has.
+    void drawHealthBars();
+    void drawHealthBar(bw::Unit unit, double ratio, bw::Color color, int yOffset);
 };
